Event_management/requirements.txt
# ============================================================
# File: requirements.txt
# Django Event Management System — All Dependencies
# ============================================================

# Core Django
Django==5.0.6
django-environ==0.11.2

# Database
psycopg2-binary==2.9.9

# Authentication
django-allauth==0.63.6

# QR Code & PDF
qrcode[pil]==7.4.2
reportlab==4.2.2
Pillow==10.3.0

# Email
django-anymail==10.3

# Payment
razorpay==1.4.2

# Task Queue
celery==5.4.0
redis==5.0.6
django-celery-results==2.5.1
django-celery-beat==2.6.0

# Static Files (Production)
whitenoise==6.7.0

# Utilities
python-slugify==8.0.4
django-crispy-forms==2.3
crispy-bootstrap5==2024.2
django-htmx==1.19.0

# Security
django-axes==6.4.0
django-csp==3.8

# Export
openpyxl==3.1.4

# Dev Tools
python-decouple==3.8
gunicorn==22.0.0
